feat: 🎸 leaderboard link optimization
